Centralize code for saving generated files
Fix bug where dryrun output method would be shown if there were no changes and verbose was not set
