+++ .changeset/slimy-cats-wait.md
@@ -1,0 +1,8 @@
+---
+'@featureboard/code-generator': patch
+'@featureboard/nx-plugin': patch
+'@featureboard/cli': patch
+---
+
+Centralize code for saving generated files
+Fix bug where dryrun output method would be shown if there were no changes and verbose was not set
