+++ libs/nx-plugin/src/executors/code-gen/executor.ts
@@ -4,10 +4,11 @@ import {
 } from '@featureboard/api-authentication'
 import {
     codeGenerator,
+    saveChanges,
     type FeatureBoardAuth,
 } from '@featureboard/code-generator'
 import { joinPathFragments, type ExecutorContext, type Tree } from '@nx/devkit'
-import { FsTree, flushChanges, printChanges } from 'nx/src/generators/tree'
+import { FsTree } from 'nx/src/generators/tree'
 import { API_ENDPOINT, CLIENT_ID } from '../../shared/config'
 import type { CodeGenExecutorSchema } from './schema'
 
@@ -66,20 +67,4 @@ export async function codeGenExecutor(
     }
 }
 
-function saveChanges(tree: Tree, dryRun: boolean, verbose: boolean) {
-    const changes = tree.listChanges()
-    if (verbose) {
-        printChanges(changes)
-        if (changes.length === 0) {
-            console.log(`\nFiles are the same no changes were made.`)
-            return
-        }
-    }
-
-    if (!dryRun && changes.length !== 0) {
-        flushChanges(tree.root, changes)
-        return
-    }
-    console.log(`\nNOTE: The "dryRun" flag means no changes were made.`)
-}
 export default codeGenExecutor
